scsi: core: Support allocating a pseudo SCSI device
Add the flag 'alloc_pseudo_sdev' to the SCSI host template and allocate a
pseudo SCSI device if the flag is set.  This device has the SCSI ID
<max_id>:U64_MAX so it won't clash with any devices the LLD might create.
It's excluded from scanning and will not show up in sysfs. Additionally,
pseudo SCSI devices are skipped by shost_for_each_device(). This prevents
that the SCSI error handler tries to submit a reset to a non-existent
logical unit.

The intention is to use this device to send internal commands to the
storage device.

Cc: John Garry <john.g.garry@oracle.com>
Signed-off-by: Hannes Reinecke <hare@suse.de>
[ bvanassche: edited patch description / renamed host_sdev into
  pseudo_sdev / unexported scsi_get_host_dev() / modified error path in
  scsi_get_pseudo_dev() / skip pseudo devices in __scsi_iterate_devices()
  and also when calling sdev_init(), sdev_configure() and sdev_destroy().
  See also
  https://lore.kernel.org/linux-scsi/20211125151048.103910-2-hare@suse.de/ ]
Signed-off-by: Bart Van Assche <bvanassche@acm.org>