locking/mutex: Change the atomic_dec_and_mutex_lock() return type
Change the return type of atomic_dec_and_mutex_lock() from int into
bool. This will make it easier to add a thread-safety annotation to
this function. This patch does not change the behavior of any kernel
code because all atomic_dec_and_mutex_lock() callers either check whether
the return value of this function is zero or that it differs from zero.

This patch makes the atomic_dec_and_mutex_lock() signature consistent
with that of refcount_dec_and_mutex_lock().

Signed-off-by: Bart Van Assche <bvanassche@acm.org>